🚚 api: move generate-results-download-link to results certification context
