+++ api/tests/certification/results/acceptance/application/certification-results-route_test.js
@@ -8,6 +8,7 @@ import {
   databaseBuilder,
   expect,
   generateValidRequestAuthorizationHeader,
+  insertUserWithRoleSuperAdmin,
 } from '../../../../test-helper.js';
 
 describe('Certification | Results | Acceptance | Application | Routes | certification results', function () {
@@ -264,6 +265,73 @@ describe('Certification | Results | Acceptance | Application | Routes | certific
 
         // then
         expect(response.statusCode).to.equal(200);
+      });
+    });
+  });
+
+  describe('GET /api/admin/sessions/{sessionId}/generate-results-download-link', function () {
+    context('when user is Super Admin', function () {
+      it('should return a 200 status code response', async function () {
+        // given
+        const sessionId = 121;
+        const options = {
+          method: 'GET',
+          url: `/api/admin/sessions/${sessionId}/generate-results-download-link`,
+          payload: {},
+        };
+        const server = await createServer();
+        await insertUserWithRoleSuperAdmin();
+        databaseBuilder.factory.buildSession({ id: 121 });
+        await databaseBuilder.commit();
+
+        // when
+        options.headers = { authorization: generateValidRequestAuthorizationHeader() };
+        const response = await server.inject(options);
+
+        // then
+        expect(response.statusCode).to.equal(200);
+      });
+    });
+
+    context('when user is not SuperAdmin', function () {
+      it('should return 403 HTTP status code', async function () {
+        // given
+        const sessionId = 121;
+        const options = {
+          method: 'GET',
+          url: `/api/admin/sessions/${sessionId}/generate-results-download-link`,
+          payload: {},
+        };
+        const server = await createServer();
+        await insertUserWithRoleSuperAdmin();
+
+        // when
+        options.headers = { authorization: generateValidRequestAuthorizationHeader(1111) };
+        const response = await server.inject(options);
+
+        // then
+        expect(response.statusCode).to.equal(403);
+      });
+    });
+
+    context('when user is not connected', function () {
+      it('should return 401 HTTP status code if user is not authenticated', async function () {
+        // given
+        const sessionId = 121;
+        const options = {
+          method: 'GET',
+          url: `/api/admin/sessions/${sessionId}/generate-results-download-link`,
+          payload: {},
+        };
+        const server = await createServer();
+        await insertUserWithRoleSuperAdmin();
+
+        // when
+        options.headers = {};
+        const response = await server.inject(options);
+
+        // then
+        expect(response.statusCode).to.equal(401);
       });
     });
   });
