+++ api/tests/certification/results/unit/application/certification-results-route_test.js
@@ -57,4 +57,32 @@ describe('Certification | Results | Unit | Application | Certification Results R
       });
     });
   });
+
+  describe('GET /api/admin/sessions/{sessionId}/generate-results-download-link', function () {
+    it('return forbidden access if user has METIER role', async function () {
+      // given
+      sinon
+        .stub(securityPreHandlers, 'hasAtLeastOneAccessOf')
+        .withArgs([
+          securityPreHandlers.checkAdminMemberHasRoleSuperAdmin,
+          securityPreHandlers.checkAdminMemberHasRoleCertif,
+          securityPreHandlers.checkAdminMemberHasRoleSupport,
+        ])
+        .callsFake(
+          () => (request, h) =>
+            h
+              .response({ errors: new Error('forbidden') })
+              .code(403)
+              .takeover(),
+        );
+      const httpTestServer = new HttpTestServer();
+      await httpTestServer.register(moduleUnderTest);
+
+      // when
+      const response = await httpTestServer.request('GET', '/api/admin/sessions/1/generate-results-download-link');
+
+      // then
+      expect(response.statusCode).to.equal(403);
+    });
+  });
 });
