+++ api/tests/identity-access-management/unit/domain/usecases/validate-user-account-email.usecase.test.js
@@ -1,4 +1,5 @@
 import { validateUserAccountEmail } from '../../../../../src/identity-access-management/domain/usecases/validate-user-account-email.usecase.js';
+import { config } from '../../../../../src/shared/config.js';
 import { monitoringTools } from '../../../../../src/shared/infrastructure/monitoring-tools.js';
 import { domainBuilder, expect, sinon } from '../../../../test-helper.js';
 
@@ -7,9 +8,11 @@ describe('Unit | Identity Access Management | Domain | UseCase | validate-user-a
   let clock, now;
 
   const token = '00000000-0000-0000-0000-000000000000';
-  const defaultRedirectionUrl = 'https://app.pix.fr/connexion';
+  let defaultRedirectionUrl;
 
   beforeEach(function () {
+    defaultRedirectionUrl = `${config.domain.pixApp + config.domain.tldFr}/connexion`;
+
     emailValidationDemandRepository = {
       get: sinon.stub(),
       remove: sinon.stub(),
