+++ api/tests/unit/application/session/index_test.js
@@ -259,34 +259,6 @@ describe('Unit | Application | Sessions | Routes', function () {
         expect(response.statusCode).to.equal(403);
       });
     });
-
-    describe('GET /api/admin/sessions/{id}/generate-results-download-link', function () {
-      it('return forbidden access if user has METIER role', async function () {
-        // given
-        sinon
-          .stub(securityPreHandlers, 'hasAtLeastOneAccessOf')
-          .withArgs([
-            securityPreHandlers.checkAdminMemberHasRoleSuperAdmin,
-            securityPreHandlers.checkAdminMemberHasRoleCertif,
-            securityPreHandlers.checkAdminMemberHasRoleSupport,
-          ])
-          .callsFake(
-            () => (request, h) =>
-              h
-                .response({ errors: new Error('forbidden') })
-                .code(403)
-                .takeover(),
-          );
-        const httpTestServer = new HttpTestServer();
-        await httpTestServer.register(moduleUnderTest);
-
-        // when
-        const response = await httpTestServer.request('GET', '/api/admin/sessions/1/generate-results-download-link');
-
-        // then
-        expect(response.statusCode).to.equal(403);
-      });
-    });
   });
 
   describe('DELETE /api/sessions/{id}', function () {
